refactor(repeaterfield): refactor to agreed format
